Siirrä raporttien käli componentit omiin tiedostoihin
