+++ web/app/raportit/Raportit.jsx
@@ -4,13 +4,9 @@ import Text from '../i18n/Text'
 import Bacon from 'baconjs'
 import OrganisaatioPicker from '../virkailija/OrganisaatioPicker'
 import Atom from 'bacon.atom'
-import DateInput from '../date/DateInput'
-import {showError} from '../util/location'
-import {formatISODate} from '../date/date'
-import {appendQueryParams} from '../util/location'
-import {generateRandomPassword} from '../util/password'
 import Http from '../util/http'
-import Dropdown from '../components/Dropdown'
+import {AikajaksoRaportti} from './AikajaksoRaportti'
+import {VuosiluokkaRaporttiPaivalta} from './VuosiluokkaRaporttiPaivalta'
 
 export const raportitContentP = () => {
   const oppilaitosAtom = Atom()
@@ -107,143 +103,4 @@ const PerusopetuksenVuosiluokka = ({oppilaitosAtom}) => {
     description={descriptionText}
     example={exampleText}
   />)
-}
-
-const AikajaksoRaportti = ({oppilaitosAtom, apiEndpoint, title, description}) => {
-  const alkuAtom = Atom()
-  const loppuAtom = Atom()
-  const submitBus = Bacon.Bus()
-
-  const password = generateRandomPassword()
-
-  const downloadExcelP = Bacon.combineWith(
-    oppilaitosAtom, alkuAtom, loppuAtom,
-    (o, a, l) => o && a && l && (l.valueOf() >= a.valueOf()) && {oppilaitosOid: o.oid, alku: formatISODate(a), loppu: formatISODate(l), password, baseUrl: `/koski/api/raportit${apiEndpoint}`}
-  )
-  const downloadExcelE = submitBus.map(downloadExcelP)
-    .flatMapLatest(downloadExcel)
-
-  downloadExcelE.onError(e => { showError(e) })
-
-  const inProgressP = submitBus.awaiting(downloadExcelE.mapError())
-  const submitEnabledP = downloadExcelP.map(x => !!x).and(inProgressP.not())
-  const buttonTextP = inProgressP.map((inProgress) => <Text name={!inProgress ? 'Lataa Excel-tiedosto' : 'Ladataan...'}/>)
-
-  return (<section>
-    <h2>{title}</h2>
-    <p>{description}</p>
-    <div className='parametri'>
-      <label><Text name='Aikajakso'/></label>
-      <div className='date-range'>
-        <DateInput value={alkuAtom.get()} valueCallback={(value) => alkuAtom.set(value)} validityCallback={(valid) => !valid && alkuAtom.set(undefined)} />
-        {' — '}
-        <DateInput value={loppuAtom.get()} valueCallback={(value) => loppuAtom.set(value)} validityCallback={(valid) => !valid && loppuAtom.set(undefined)} />
-      </div>
-    </div>
-    <div className='password'><Text name='Excel-tiedosto on suojattu salasanalla'/> {password}</div>
-    <button className='koski-button' disabled={submitEnabledP.not()} onClick={e => { e.preventDefault(); submitBus.push(); return false }}>{buttonTextP}</button>
-  </section>)
-}
-
-const VuosiluokkaRaporttiPaivalta = ({oppilaitosAtom, apiEndpoint, title, description, example}) => {
-  const paivaAtom = Atom()
-  const vuosiluokkaAtom = Atom('')
-  const submitBus = Bacon.Bus()
-
-  const password = generateRandomPassword()
-
-  const downloadExcelP = Bacon.combineWith(
-    oppilaitosAtom, paivaAtom, vuosiluokkaAtom,
-    (o, p, v) => o && p && v && ({oppilaitosOid: o.oid, paiva: formatISODate(p), vuosiluokka:(v), password, baseUrl: `/koski/api/raportit${apiEndpoint}`})
-  )
-  const downloadExcelE = submitBus.map(downloadExcelP)
-    .flatMapLatest(downloadExcel)
-
-  downloadExcelE.onError(e => { showError(e) })
-
-  const inProgressP = submitBus.awaiting(downloadExcelE.mapError())
-  const submitEnabledP = downloadExcelP.map(x => !!x).and(inProgressP.not())
-  const buttonTextP = inProgressP.map((inProgress) => <Text name={!inProgress ? 'Lataa Excel-tiedosto' : 'Ladataan...'}/>)
-  const vuosiluokat = [1, 2, 3, 4, 5, 6, 7, 8, 10]
-
-  return (<section>
-    <h2>{title}</h2>
-    <p>{description}</p>
-    <p>{example}</p>
-    <div>
-      <div className='parametri'>
-        <label><Text name='Päivä'/></label>
-        <DateInput value={paivaAtom.get()} valueCallback={(value) => paivaAtom.set(value)} validityCallback={(valid) => !valid && paivaAtom.set(undefined)} />
-      </div>
-    </div>
-    <div className='dropdown-selection parametri'>
-      <label><Text name='Vuosiluokka'/></label>
-      <VuosiluokkaDropdown value={vuosiluokkaAtom} vuosiluokat={vuosiluokat}/>
-    </div>
-    <div className='password'><Text name='Excel-tiedosto on suojattu salasanalla'/> {password}</div>
-    <button className='koski-button' disabled={submitEnabledP.not()} onClick={e => { e.preventDefault(); submitBus.push(); return false }}>{buttonTextP}</button>
-  </section>)
-}
-
-const VuosiluokkaDropdown = ({value, vuosiluokat}) => (
-  <div>
-    {value.map(v => (
-      <Dropdown
-        options={vuosiluokat}
-        keyValue={(key) => key}
-        displayValue={(dVal) => dVal === 10 ? 'Peruskoulun päättävät' : dVal}
-        selected={v}
-        onSelectionChanged={(input) => value.set(input)}
-      />))}
-  </div>
-)
-
-const downloadExcel = (params) => {
-  let iframe = document.getElementById('raportti-iframe')
-  if (iframe) {
-    iframe.parentNode.removeChild(iframe)
-  }
-  iframe = document.createElement('iframe')
-  iframe.id = 'raportti-iframe'
-  iframe.style.display = 'none'
-  document.body.appendChild(iframe)
-
-  const resultBus = Bacon.Bus()
-  let downloadTimer
-
-  iframe.addEventListener('load', () => {
-    var response
-    try {
-      response = { text: iframe.contentDocument.body.textContent, httpStatus: 400 }
-    } catch (err) {
-      response = { text: 'Tuntematon virhe', httpStatus: 500 }
-    }
-    window.clearInterval(downloadTimer)
-    resultBus.error(response)
-    resultBus.end()
-  })
-
-  const downloadToken = 'raportti' + new Date().getTime()
-  const {baseUrl, ...queryParams} = params
-  const url = appendQueryParams(baseUrl, {...queryParams, downloadToken})
-  iframe.src = url
-
-  // detect when download has started by polling a cookie set by the backend.
-  // based on https://stackoverflow.com/questions/1106377/detect-when-browser-receives-file-download
-  let attempts = 360
-  downloadTimer = window.setInterval(() => {
-    if (document.cookie.indexOf('koskiDownloadToken=' + downloadToken) >= 0) {
-      window.clearInterval(downloadTimer)
-      resultBus.push(true)
-      resultBus.end()
-    } else {
-      if (--attempts < 0) {
-        window.clearInterval(downloadTimer)
-        resultBus.error({text: 'Timeout', httpStatus: 400})
-        resultBus.end()
-      }
-    }
-  }, 1000)
-
-  return resultBus
 }
